+++ coordinator/core/ssh.go
@@ -8,11 +8,11 @@ import (
 
 // CreateSSHUsers creates UNIX users with respective SSH access on the system the coordinator is running on when defined in the config.
 func (c *Core) CreateSSHUsers(sshUserKeys []ssh.UserKey) error {
-	sshAccess := ssh.NewSSHAccess(c.linuxUserManager)
+	sshAccess := ssh.NewAccess(c.linuxUserManager)
 	ctx := context.Background()
 
 	for _, pair := range sshUserKeys {
-		if err := sshAccess.DeploySSHAuthorizedKey(ctx, pair); err != nil {
+		if err := sshAccess.DeployAuthorizedKey(ctx, pair); err != nil {
 			return err
 		}
 	}